requirements.txt
django==4.2
asgiref==3.8.1
tzdata==2024.1
sqlparse==0.4.4
djangorestframework==3.14.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9